Add provider and distro options (#80)
The previous option of platform has been split into two separate options,
provider and distro. This allows us to distinguish between cloud-managed
clusters and user-managed clusters.

The reason " " (with a space) is allowed as an option is to work around UI
wizard that is not able to omit the option or provide just an empty space.