Add optional native OTel logs collection (#197)
* Add optional native OTel logs collection

We want to replace fluentd logs collection with native OTel logs collection. For now we adding this as an option. Later it will become the default way to collect logs

The configuration for native logs collection mostly borrowed from https://github.com/splunk/sck-otel

Extra changes to the borrowed code:
- move "stream" from resource to log attribute
- fix otel agent container path: otelcollector -> otel-collector