Remove temporary initContainer, helper for OTel native log path move (#358)
This initContainer was needed to support change of the default path for native logs collection checkpointing to avoid data duplication with upgrade. We waited enough to remove the helper. Also it relies on otel-collector-image with smartagent bundle which is not available for ARM architectures. So in some cases, this initContainer can block the collector from running on ARM architectures.