Fix log collection in GKE Autopilot with fluentd
This change starts the transition towards assuming that containerd is the container engine used by default instead of trying to detect it. We are confident that at least GKE Autopilot don't use the docker engine anymore so we can exclude the autodection for that distribution. We will be adding other distributions to the exclusion list going forward, and eventually we can drop support of docker engine.