Remove redundant gateway-only example
We have two similar gateway-only examples. Given that we have to many examples now, it should be good to consider one of the gateway-only examples redundant and delete it.

We pick the one with overridden resources which required a fix: it used 4Gb for memory limit which wasn't properly translated to SPLUNK_MEMORY_TOTAL_MIB.