Fix `k8s.cluster.name` resource attribute for logs in GCP
Following the merge of https://github.com/signalfx/splunk-otel-collector-chart/pull/660, the `resource` processor in the logs pipeline was placed before the `resourcedetection` processor. In contrast, the other pipelines have the `resource` processor after the `resourcedetection` processor. This order is crucial in GCP because the `resourcedetection` processor's `gcp` detector sets the `k8s.cluster.name` attribute, which we then override with the user-provided value in the `resource` processor. By keeping the `resource` processor after the `resourcedetection` processor, we ensure that any `k8s.cluster.name` attribute is always set with the user-provided value.

This change was implemented to ensure that the order of processors is consistent across all pipelines, guaranteeing that the `k8s.cluster.name` attribute is always set to the value provided in the `clusterName` field of the values.yaml file.