Set cluster_name for host logs too if renameFieldsSck is enabled
Up to now the cluster_name field was only set in `resource/logs`. This
snippet was only included for the container logs, but not for host logs
(files, journald).

This change moves the cluster_name field to the `resource` snippet,
which is according inline comment used for all things that pass through
otel.

Fixes #682