Disable opencensus.resourcetype attribute in cluster_receiver (#914)
Disable a redundant `opencensus.resourcetype` resource attribute in k8s_cluster receiver. The attribute will be removed from the receiver upstream anyway, but this will happen in a deprecation process, so we need to put this config in the receiver to avoid throwing deprecation warnings.

This change doe not affect Splunk Observability users, because it has been already disabled in the signalfx exporter default translation rules.

This can be a breaking change for Splunk Platform users relying on this attribute. If required, it can added back by configuring an additional `resource` processor.

```
clusterReceiver:
  config:
    processors:
      resource/add_opencensus_resourcetype:
        attributes:
        - action: insert
          key: opencensus.resourcetype
          value: k8s
    service:
      pipelines:
        metrics:
          processors: 
            - memory_limiter
            - batch
            - resource
            - resource/k8s_cluster
            - resource/add_opencensus_resourcetype
```