Disable signalfx exporter default translations in clusterReceiver
This change disables signalfx exporter default translations the clusterReceiver deploymentto avoid wasting CPU cycles. Removing of `opencensus.resourcetype` attribute was the only translation applicable to k8s_cluster receiver metrics. After it the attribute was disabled on the receiver side in https://github.com/signalfx/splunk-otel-collector-chart/pull/914, the default translations are not needed anymore.

This can be a breaking change for Splunk Observability users who use the deprecated `translation_rules` option in signalfx exporter in clusterReceiver. The option has been deprecated for more than half a year, so we should encourage users to not use it anymore.