add original names of subcharts to values schema
values schema file contains entries for sub-charts but it uses the alias as the key. This works when using splunk-otel-collector-chart directly but when the chart is itself used as a sub-chart, helm fails to lint the chart resulting in issues in CI and deployment pipelines.

This commit explicitly allows both alias and original names of sub-charts in the values schema to work around the limitation in helm.

This should be ideally fixed in helm but this patch is minor enough with a relative big amount of impact so I think we should carry it. 

fixes https://github.com/signalfx/splunk-otel-collector-chart/issues/874