Enable validation of `enable` field
Since `additionalProperties: false` in the `values.schema.json` file, it is not possible to use common pattern to include `splunk-otel-collector` as a subchart using `condition` field in the parent charts `charts.yaml` (as explained in the helm documentation https://helm.sh/docs/topics/charts/#tags-and-condition-fields-in-dependencies). 

This PR adds a validation for `enabled` boolean field to solve the issue.