[chore] Update collector image to 0.120.0
This also updates a test validating in expected internal collector metrics. The upgrade to Prometheus 3.0 caused that `.` in metric names are not replaced to `_` in the prometheus receiver. As the result, there is one metric affected in the test `otelcol_processor_filter_logs_filtered` updated to `otelcol_processor_filter_logs.filtered`. Looks like there are few of possible metrics affected. Not a big problem. I'll update the changelog upstream with more details, then we can pick it up.