[chore] Remove required section from values.schema.yaml
The requirement isn't applicable to the regular helm install/upgrade since the `values.yaml` have empty string values by defaults for the `cloudProvider` and `distribution` fields. However, the EKS Addon validator treats the `required` field differently making the `cloudProvider` and `distribution` to be required from the user with every installation even if we pre-set `aws` and `eks` in the values.yaml.